% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/EFSlabels.R
\name{EFSlabels}
\alias{EFSlabels}
\title{Label data exported from unipark.}
\usage{
EFSlabels(dataFileLocation, codebookFileLocation)
}
\arguments{
\item{dataFileLocation}{A .csv file exported from unipark.}

\item{codebookFileLocation}{The codebook file exported from unipark (.csv with UTF-8, NOT from 'Export variable names for external editing').}
}
\value{
A data frame \code{dataFileLocation} with labels from \code{codebookFileLocation}.
}
\description{
\code{EFSlabels} returns a data frame with labeled values for each variable found in the codebook.
}
\details{
The parameter \code{dataFileLocation} should point to the location of the .csv file on your drive.
\code{codebookFileLocation} should point to the location of the .csv codebook file.
Export this file from unipark by selecting Documentation > Codebook > Export > Export as CSV.
}
\examples{
df <- EFSlabels("data_project_206763_2015_09_23.csv","codebook_project_206763_2015_09_23.csv")
head(df)
}

